modulize owlet powlet and dom
